+++ src/containerapp/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 0.3.11
 ++++++
+* Add keda scale rule parameters to 'az containerapp create', 'az containerapp update' and 'az containerapp revision copy'
 * Add new dapr params to 'az containerapp dapr enable' and 'az containerapp create'
 * 'az containerapp up': autogenerate a docker container with --source when no dockerfile present